requirements.txt
anthropic>=0.39.0
flask>=3.0.0
gunicorn>=21.2.0
twilio>=9.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0
apscheduler>=3.10.4
requests>=2.31.0
